More populate files clean up
