+++ packages/sdc-populate/src/index.ts
@@ -15,22 +15,7 @@
  * limitations under the License.
  */
 
-import type { PopulateInputParameters } from './Interfaces';
-import type { OperationOutcome, Parameters, Reference } from 'fhir/r4';
-import { OperationOutcomeIssue } from 'fhir/r4';
-import { isQuestionnaireDataParameter, isSubjectParameter } from './TypePredicates';
-import type { InputParameters } from './interfaces/inputParameters.interface';
-import { FetchResourceCallback } from './interfaces/callback.interface';
-import { fetchQuestionnaire } from './getQuestionnaire';
-import { getContextMap } from './getContexts';
-import { readPopulationExpressions } from './ReadPopulationExpressions';
-import { sortResourceArrays } from './SortBundles';
-import { constructResponse } from './ConstructQuestionnaireResponse';
-import { evaluateItemPopulationContexts } from './EvaluateItemPopulationContexts';
-import { evaluateInitialExpressions } from './EvaluateInitialExpressions';
-import { createOutputParameters } from './CreateOutputParameters';
-import type { OutputParameters } from './interfaces/outputParameters.interface';
-
+// Interfaces
 export { FetchResourceCallback } from './interfaces/callback.interface';
 export {
   IdentifierParameter,
@@ -38,76 +23,8 @@ export {
 } from './interfaces/inputParameters.interface';
 export { IssuesParameter, ResponseParameter } from './interfaces/outputParameters.interface';
 
-/**
- * Main function of this populate module.
- * Input and output specific parameters conformant to the SDC populate specification.
- *
- * @author Sean Fong
- */
-export default async function populate(
-  parameters: InputParameters,
-  fetchResourceCallback: FetchResourceCallback,
-  requestConfig: any
-): Promise<OutputParameters | OperationOutcome> {
-  const issues: OperationOutcomeIssue[] = [];
+// Type Predicates
+export { isInputParameters } from './typePredicates';
 
-  const questionnaire = await fetchQuestionnaire(parameters, fetchResourceCallback, requestConfig);
-  if (questionnaire.resourceType === 'OperationOutcome') {
-    return questionnaire;
-  }
-
-  const subjectReference = parameters.parameter.find((param) => isSubjectParameter(param))
-    ?.valueReference as Reference;
-
-  // Create contextMap to hold variables for population
-  let contextMap = await getContextMap(
-    parameters,
-    questionnaire,
-    fetchResourceCallback,
-    requestConfig,
-    issues
-  );
-
-  // Read expressions to be populated from questionnaire recursively
-  // i.e. itemPopulationContext, initialExpression
-  const populationExpressions = readPopulationExpressions(questionnaire);
-
-  // Evaluate itemPopulationContexts and add them to contextMap
-  contextMap = evaluateItemPopulationContexts(
-    populationExpressions.itemPopulationContexts,
-    contextMap,
-    issues
-  );
-  contextMap = sortResourceArrays(contextMap);
-
-  // Evaluate initialExpressions
-  const evaluatedInitialExpressions = evaluateInitialExpressions(
-    populationExpressions.initialExpressions,
-    contextMap,
-    issues
-  );
-
-  // Construct response from initialExpressions
-  const questionnaireResponse = await constructResponse(
-    questionnaire,
-    subjectReference,
-    evaluatedInitialExpressions
-  );
-
-  return createOutputParameters(questionnaireResponse, issues);
-}
-
-/**
- * Checks if the parameters passed satisfies the conditions of populateInputParameters.
- *
- * @author Sean Fong
- */
-export function isPopulateInputParameters(
-  parameters: Parameters
-): parameters is PopulateInputParameters {
-  const questionnairePresent = !!parameters.parameter?.find(isQuestionnaireDataParameter);
-
-  const subjectPresent = !!parameters.parameter?.find(isSubjectParameter);
-
-  return questionnairePresent && subjectPresent;
-}
+// Functions
+export { populate } from './populate';
