+++ packages/sdc-populate/src/processValueSets.ts
@@ -22,7 +22,7 @@ import type {
   QuestionnaireResponseItemAnswer,
   ValueSet
 } from 'fhir/r4';
-import type { ValueSetPromise } from './Interfaces';
+import type { ValueSetPromise } from './interfaces/expressions.interface';
 import * as FHIR from 'fhirclient';
 
 const ONTOSERVER_ENDPOINT = 'https://r4.ontoserver.csiro.au/fhir/';
